backend/main.py
from fastapi import FastAPI
from pydantic import BaseModel
from backend.rag import ask_question
from fastapi.middleware.cors import CORSMiddleware

app = FastAPI()

# ✅ ADD THIS
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],  # allow all (good for testing)
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

class Question(BaseModel):
    question: str

@app.post("/ask")
def ask(q: Question):
    answer = ask_question(q.question)
    return {"answer": answer}