requirements.txt
fastapi
uvicorn
streamlit
faiss-cpu
openai
pypdf
python-dotenv
numpy
